+++ apps/lite/ui/src/routes/project/$id/workspace/OperationTarget.tsx
@@ -1,7 +1,6 @@
 import { operandEquals, type Operand } from "#ui/operands.ts";
 import { parseDragData } from "./DragData.ts";
 import styles from "./OperationTarget.module.css";
-import { OperationTooltip } from "./OperationTooltip.tsx";
 import {
 	getOperation,
 	getOperations,
@@ -16,9 +15,58 @@ import {
 	attachInstruction,
 	extractInstruction,
 } from "@atlaskit/pragmatic-drag-and-drop-hitbox/list-item";
-import { mergeProps, useRender } from "@base-ui/react";
+import { mergeProps, Tooltip, useRender } from "@base-ui/react";
 import { Match, pipe } from "effect";
 import { FC, useEffect, useEffectEvent, useRef } from "react";
+import { TooltipPopup } from "#ui/components/Tooltip.tsx";
+
+const OperationTooltip: FC<
+	{
+		target: Operand;
+		isActive: boolean;
+		projectId: string;
+	} & useRender.ComponentProps<"div">
+> = ({ target, isActive, projectId, render, ...props }) => {
+	const outlineMode = useAppSelector((state) => selectProjectOutlineModeState(state, projectId));
+
+	const tooltip = isActive
+		? Match.value(outlineMode).pipe(
+				Match.when({ _tag: "Absorb" }, () => <>Absorb target</>),
+				Match.when({ _tag: "Transfer", value: { _tag: "Pointer" } }, ({ value: mode }) =>
+					mode.target && mode.operationType !== null
+						? getOperation({
+								source: mode.source,
+								target: mode.target,
+								operationType: mode.operationType,
+							})?.label
+						: null,
+				),
+				Match.when(
+					{ _tag: "Transfer", value: { _tag: "Keyboard" } },
+					({ value: mode }) =>
+						getOperation({
+							source: mode.source,
+							target,
+							operationType: mode.operationType,
+						})?.label,
+				),
+				Match.orElse(() => null),
+			)
+		: null;
+
+	const trigger = useRender({ render, props });
+
+	return (
+		<Tooltip.Root open={tooltip != null} disableHoverablePopup>
+			<Tooltip.Trigger render={trigger} />
+			<Tooltip.Portal>
+				<Tooltip.Positioner sideOffset={8} side="right">
+					<Tooltip.Popup render={<TooltipPopup />}>{tooltip}</Tooltip.Popup>
+				</Tooltip.Positioner>
+			</Tooltip.Portal>
+		</Tooltip.Root>
+	);
+};
 
 type DropTargetParams = Parameters<typeof dropTargetForElements>[0];
 type GetDataArgs = Parameters<NonNullable<DropTargetParams["getData"]>>[0];
